Move plotting classes into seperate function
